+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Changelog
 
-1.0.0 (2020-02-03)
+1.1.0 (2020-06-18)
+- Add override to settings to disable automatic frontend display
+- Add yml config and template variable to enable developer repositioning
+
+1.0.0 (2020-05-26)
 - Initial release
+
+(2020-02-03)
+- Initial development